requirements.txt
fastapi
uvicorn[standard]
aiortc==1.13
PyJWT
